requirements.txt
google-genai
python-dotenv
pandas
streamlit